% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{authorize_api_key}
\alias{authorize_api_key}
\title{Authorize the Eventbrite personal key}
\usage{
authorize_api_key(token)
}
\arguments{
\item{token}{The access token}
}
\description{
Authorize the Eventbrite personal key
}
