% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_api_response}
\alias{get_api_response}
\title{Authorize the Eventbrite personal key}
\usage{
get_api_response(token, ...)
}
\arguments{
\item{token}{The access token}
}
\description{
Authorize the Eventbrite personal key
}
