% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{read_event_data}
\alias{read_event_data}
\title{Read Event Data}
\usage{
read_event_data(eventid)
}
\arguments{
\item{eventid}{Character vector of length \code{1L}; The ID of the given
event whose data we want to poll.}
}
\description{
Reads data on events from the server
}
