% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{setup_eventbrite_oauth}
\alias{setup_eventbrite_oauth}
\title{Set up OAuth 2.0 for the Eventbrite API}
\usage{
setup_eventbrite_oauth(api.key = NULL, client.secret = NULL, ...)
}
\arguments{
\item{api.key}{The application's \emph{API Key}.}

\item{client.secret}{The application's \emph{Client Secret}.}

\item{...}{Arguments passed to \code{\link[httr]{oauth2.0_token}}.}
}
\value{
An R6 object of type 'Token'.
}
\description{
Will set up OAuth for the user, with the package playing the role of 'client'.
}
